feat(add-to/cart): Add option for showing it when the product is unavailable
If the button is shown in some tiles and in some not, and if the button affects the dimensions of the tile
(e.g. when it's shown as a additional block in the bottom of the tile, extending its' height),
it will cause some tiles to be taller than the others.

This commit allows the dev to enforce the button to be shown for all products (including unavailable ones)
by setting an option in xml (and styling it differently through css, when the product is unavailable).